+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 #### 功能构建
 
+- 新增 MapperRegistry。
+  - com.dwarfeng.fdr.impl.handler.mapper.IdentityMapperRegistry
 - 优化 Filter, Trigger, Mapper 的代码结构。
   - com.dwarfeng.fdr.impl.handler.filter.AbstractFilterRegistry
   - com.dwarfeng.fdr.impl.handler.trigger.AbstractTriggerRegistry